refactor(views): group screens and shared parts into directories
