+++ crates/views/src/lib.rs
@@ -1,28 +1,15 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-mod adaptive;
-mod artist;
-mod cells;
-mod detail;
-mod hero;
-mod home;
-mod library;
-mod login;
-mod page;
-mod quick_picks;
-mod release_card;
 mod root;
-mod search;
-mod settings;
-mod song;
-mod tracks;
+mod screens;
+mod shared;
 
-use adaptive::Adaptive;
-use artist::ArtistView;
-use detail::DetailView;
-use home::HomeView;
-pub use library::LibraryView;
-pub use login::LoginView;
 pub use root::Root;
-pub use settings::SettingsView;
-use song::SongView;
+use screens::artist::ArtistView;
+use screens::detail::DetailView;
+use screens::home::HomeView;
+pub use screens::library::LibraryView;
+pub use screens::login::LoginView;
+pub use screens::settings::SettingsView;
+use screens::song::SongView;
+use shared::adaptive::Adaptive;
