+++ crates/views/src/screens/artist.rs
@@ -13,10 +13,10 @@ use ui::ActiveTheme as _;
 use ui::{Button, ColumnSpec, GridDelegate, GridEvent, GridState, Scrollbar, Scroller, Text, grid};
 use workspace::Chrome;
 
-use crate::hero::{HeroPlayButton, PageHero};
-use crate::page;
-use crate::release_card::ReleaseCard;
-use crate::tracks::{PlaybackStatus, TrackField, TrackSource, Tracks, playback_status};
+use crate::shared::hero::{HeroPlayButton, PageHero};
+use crate::shared::page;
+use crate::shared::release_card::ReleaseCard;
+use crate::shared::tracks::{PlaybackStatus, TrackField, TrackSource, Tracks, playback_status};
 
 #[derive(Clone, Copy, PartialEq, Eq)]
 enum ReleaseFilter {
