+++ crates/views/src/screens/mod.rs
@@ -1,0 +1,10 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+pub(crate) mod artist;
+pub(crate) mod detail;
+pub(crate) mod home;
+pub(crate) mod library;
+pub(crate) mod login;
+pub(crate) mod search;
+pub(crate) mod settings;
+pub(crate) mod song;
