+++ crates/views/src/shared/adaptive.rs
