+++ crates/views/src/shared/hero.rs
