+++ crates/views/src/shared/mod.rs
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+pub(crate) mod adaptive;
+pub(crate) mod cells;
+pub(crate) mod hero;
+pub(crate) mod page;
+pub(crate) mod release_card;
+pub(crate) mod tracks;
